+++ sdk/eventgrid/eventgrid/src/consumer.ts
@@ -3,12 +3,7 @@
 
 import { createSerializer } from "@azure/core-client";
 import { CloudEvent as WireCloudEvent } from "./generated/models";
-import {
-  CustomEventDataDeserializer,
-  CloudEvent,
-  EventGridEvent,
-  cloudEventReservedPropertyNames
-} from "./models";
+import { CloudEvent, EventGridEvent, cloudEventReservedPropertyNames } from "./models";
 import {
   EventGridEvent as EventGridEventMapper,
   CloudEvent as CloudEventMapper
@@ -19,37 +14,19 @@ import { systemDeserializers } from "./systemEventDecoders";
 const serializer = createSerializer();
 
 /**
- * Options for the Event Grid Consumer
- */
-export interface EventGridConsumerOptions {
-  /**
-   * Custom deserializers to use when decoding a specific event's data, based on the type
-   * field of the event.
-   */
-  customDeserializers: Record<string, CustomEventDataDeserializer>;
-}
-
-/**
- * EventGridConsumer is used to aid in processing events delivered by EventGrid. It can deserialize a JSON encoded payload
+ * EventGridDeserializer is used to aid in processing events delivered by EventGrid. It can deserialize a JSON encoded payload
  * of either a single event or batch of events as well as be used to convert the result of `JSON.parse` into an
  * `EventGridEvent` or `CloudEvent` like object.
  *
- * Unlike normal JSON deseralization, EventGridConsumer does some additional conversions:
+ * Unlike normal JSON deseralization, EventGridDeserializer does some additional conversions:
  *
  * - The consumer parses the event time property into a `Date` object, for ease of use.
  * - When deserializing an event in the CloudEvent schema, if the event contains binary data, it is base64 decoded
  *   and returned as an instance of the `Uint8Array` type.
  * - The `data` payload from system events is converted to match the interfaces this library defines.
- *
- * When constructing an `EventGridConsumer`, a map of event types to custom deserializers may be provided. When
- * deserializing, if a custom deserializer has been registered for a given event type, it will be called with the
- * data object. The object this deserializer returns will replace the existing data object.
  */
-export class EventGridConsumer {
-  readonly customDeserializers: Record<string, CustomEventDataDeserializer>;
-  constructor(options?: EventGridConsumerOptions) {
-    this.customDeserializers = options?.customDeserializers ?? {};
-  }
+export class EventGridDeserializer {
+  constructor() {}
 
   /**
    * Deserializes events encoded in the Event Grid schema.
@@ -83,10 +60,6 @@ export class EventGridConsumer {
 
       if (systemDeserializers[deserialized.eventType]) {
         deserialized.data = await systemDeserializers[deserialized.eventType](deserialized.data);
-      } else if (this.customDeserializers[deserialized.eventType]) {
-        deserialized.data = await this.customDeserializers[deserialized.eventType](
-          deserialized.data
-        );
       }
 
       events.push(deserialized as EventGridEvent<unknown>);
@@ -166,8 +139,6 @@ export class EventGridConsumer {
       // If a decoder is registered, apply it to the data.
       if (systemDeserializers[modelEvent.type]) {
         modelEvent.data = await systemDeserializers[modelEvent.type](modelEvent.data);
-      } else if (this.customDeserializers[modelEvent.type]) {
-        modelEvent.data = await this.customDeserializers[modelEvent.type](modelEvent.data);
       }
 
       // Build the "extensionsAttributes" property bag by removing all known top level properties.