+++ sdk/eventgrid/eventgrid/src/eventGridClient.ts
@@ -30,22 +30,36 @@ export type EventGridPublisherClientOptions = PipelineOptions;
 /**
  * Options for the send events operation.
  */
-export type SendEventsOptions = OperationOptions;
+export type SendEventOptions = OperationOptions;
 
 /**
- * Options for the send cloud events operation.
+ * A map of input schema names to shapes of the input for the send method on EventGridPublisherClient.
  */
-export type SendCloudEventsOptions = OperationOptions;
+export interface InputSchemaToInputTypeMap {
+  /**
+   * The shape of the input to `send` when the client is configured to send events using the Event Grid schema.
+   */
+  EventGrid: SendEventGridEventInput<unknown>;
+  /**
+   * The shape of the input to `send` when the client is configured to send events using the Cloud Event schema.
+   */
+  CloudEvent: SendCloudEventInput<unknown>;
+  /**
+   * The shape of the input to `send` when the client is configured to send events using a custom schema.
+   */
+
+  Custom: object;
+}
 
 /**
- * Options for the send custom schema events operation.
+ * Allowed schema types, to be used when constructing the EventGridPublisherClient.
  */
-export type SendCustomSchemaEventsOptions = OperationOptions;
+export type InputSchema = keyof InputSchemaToInputTypeMap;
 
 /**
  * Client class for publishing events to the Event Grid Service.
  */
-export class EventGridPublisherClient {
+export class EventGridPublisherClient<T extends InputSchema> {
   /**
    * The URL to the Event Grid endpoint.
    */
@@ -62,7 +76,12 @@ export class EventGridPublisherClient {
   private readonly client: GeneratedClient;
 
   /**
-   * Creates an instance of EventGridPublisherClient.
+   * The schema that will be used when sending events.
+   */
+  private readonly inputSchema: InputSchema;
+
+  /**
+   * Creates an instance of EventGridPublisherClient which sends events using the Event Grid Schema.
    *
    * Example usage:
    * ```ts
@@ -70,20 +89,24 @@ export class EventGridPublisherClient {
    *
    * const client = new EventGridPublisherClient(
    *    "<service endpoint>",
+   *    "EventGrid",
    *    new AzureKeyCredential("<api key>")
    * );
    * ```
    *
-   * @param endpointUrl - The URL to the EventGrid endpoint, e.g. https://eg-topic.westus2-1.eventgrid.azure.net/api/events
+   * @param endpointUrl - The URL to the Event Grid endpoint, e.g. https://eg-topic.westus2-1.eventgrid.azure.net/api/events.
+   * @param inputSchema - The schema that the Event Grid endpoint is configured to accept. One of "EventGrid", "CloudEvent", or "Custom".
    * @param credential - Used to authenticate requests to the service.
-   * @param options - Used to configure the Event Grid Client
+   * @param options - Used to configure the Event Grid Client.
    */
   constructor(
     endpointUrl: string,
+    inputSchema: T,
     credential: KeyCredential | SASCredential,
     options: EventGridPublisherClientOptions = {}
   ) {
     this.endpointUrl = endpointUrl;
+    this.inputSchema = inputSchema;
 
     const libInfo = `azsdk-js-eventgrid/${SDK_VERSION}`;
     const pipelineOptions = { ...options };
@@ -108,81 +131,43 @@ export class EventGridPublisherClient {
   }
 
   /**
-   * Publishes events in the Event Grid schema. The topic must be configured to expect events in the Event Grid schema.
+   * Sends events to a topic.
    *
-   * @param message - One or more events to publish
+   * @param events - The events to send. The events should be in the schema used when constructing the client.
+   * @param options - Options to control the underlying operation.
    */
-  async sendEvents(
-    events: SendEventGridEventInput<any>[],
-    options?: SendEventsOptions
-  ): Promise<void> {
-    const { span, updatedOptions } = createSpan(
-      "EventGridPublisherClient-sendEvents",
-      options || {}
-    );
+  async send(events: InputSchemaToInputTypeMap[T][], options?: SendEventOptions) {
+    const { span, updatedOptions } = createSpan("EventGridPublisherClient-send", options || {});
 
     try {
-      return await this.client.publishEvents(
-        this.endpointUrl,
-        (events || []).map(convertEventGridEventToModelType),
-        updatedOptions
-      );
-    } catch (e) {
-      span.setStatus({ code: CanonicalCode.UNKNOWN, message: e.message });
-      throw e;
-    } finally {
-      span.end();
-    }
-  }
-
-  /**
-   * Publishes events in the Cloud Events 1.0 schema. The topic must be configured to expect events in the Cloud Events 1.0 schema.
-   *
-   * @param message - One or more events to publish
-   */
-  async sendCloudEvents(
-    events: SendCloudEventInput<any>[],
-    options?: SendCloudEventsOptions
-  ): Promise<void> {
-    const { span, updatedOptions } = createSpan(
-      "EventGridPublisherClient-sendCloudEvents",
-      options || {}
-    );
-
-    try {
-      return await this.client.publishCloudEventEvents(
-        this.endpointUrl,
-        (events || []).map(convertCloudEventToModelType),
-        updatedOptions
-      );
-    } catch (e) {
-      span.setStatus({ code: CanonicalCode.UNKNOWN, message: e.message });
-      throw e;
-    } finally {
-      span.end();
-    }
-  }
-
-  /**
-   * Publishes events written using a custom schema. The topic must be configured to expect events in a custom schema.
-   *
-   * @param message - One or more events to publish
-   */
-  async sendCustomSchemaEvents(
-    events: Record<string, any>[],
-    options?: SendCustomSchemaEventsOptions
-  ): Promise<void> {
-    const { span, updatedOptions } = createSpan(
-      "EventGridPublisherClient-sendCustomSchemaEvents",
-      options || {}
-    );
-
-    try {
-      return await this.client.publishCustomEventEvents(
-        this.endpointUrl,
-        events || [],
-        updatedOptions
-      );
+      switch (this.inputSchema) {
+        case "EventGrid": {
+          return await this.client.publishEvents(
+            this.endpointUrl,
+            (events as InputSchemaToInputTypeMap["EventGrid"][]).map(
+              convertEventGridEventToModelType
+            ),
+            updatedOptions
+          );
+        }
+        case "CloudEvent": {
+          return await this.client.publishCloudEventEvents(
+            this.endpointUrl,
+            (events as InputSchemaToInputTypeMap["CloudEvent"][]).map(convertCloudEventToModelType),
+            updatedOptions
+          );
+        }
+        case "Custom": {
+          return await this.client.publishCustomEventEvents(
+            this.endpointUrl,
+            events as InputSchemaToInputTypeMap["Custom"][],
+            updatedOptions
+          );
+        }
+        default: {
+          throw new Error(`Unknown input schema type '${this.inputSchema}'`);
+        }
+      }
     } catch (e) {
       span.setStatus({ code: CanonicalCode.UNKNOWN, message: e.message });
       throw e;